# Fix low resolution cache