--- a/Engine/DataFeeds/TextSubscriptionDataSourceReader.cs
+++ b/Engine/DataFeeds/TextSubscriptionDataSourceReader.cs
@@ -40,16 +40,16 @@ namespace QuantConnect.Lean.Engine.DataFeeds
         private BaseData _factory;
         private bool _shouldCacheDataPoints;
         private static readonly MemoryCache BaseDataSourceCache = new MemoryCache("BaseDataSourceCache",
-            new NameValueCollection(10)
+            new NameValueCollection
             {
-                { "CacheMemoryLimitMegabytes", "100" },
+                { "CacheMemoryLimitMegabytes", "250" },
                 { "PhysicalMemoryLimitPercentage", "10" },
-                { "PollingInterval", TimeSpan.FromMilliseconds(10000).ToString() }
+                { "PollingInterval", TimeSpan.FromMilliseconds(15000).ToString() }
             });
         private static readonly CacheItemPolicy CachePolicy = new CacheItemPolicy
         {
             // Cache entry should be evicted if it has not been accessed in given span of time:
-            SlidingExpiration = TimeSpan.FromMinutes(5)
+            SlidingExpiration = TimeSpan.FromMinutes(1)
         };
 
         /// <summary>