--- a/Engine/DataFeeds/TextSubscriptionDataSourceReader.cs
+++ b/Engine/DataFeeds/TextSubscriptionDataSourceReader.cs
@@ -20,10 +20,9 @@ using QuantConnect.Data;
 using QuantConnect.Interfaces;
 using QuantConnect.Data.Market;
 using System.Collections.Generic;
-using System.Collections.Concurrent;
-using System.Threading;
 using QuantConnect.Data.Fundamental;
 using QuantConnect.Data.UniverseSelection;
+using QuantConnect.Logging;
 
 namespace QuantConnect.Lean.Engine.DataFeeds
 {
@@ -41,8 +40,8 @@ namespace QuantConnect.Lean.Engine.DataFeeds
         private bool _shouldCacheDataPoints;
 
         private static int CacheSize = 100;
-        private static volatile Dictionary<string, List<BaseData>> BaseDataSourceCache = new Dictionary<string, List<BaseData>>();
-        private static ConcurrentQueue<string> CacheKeys = new ConcurrentQueue<string>();
+        private static volatile Dictionary<string, List<BaseData>> BaseDataSourceCache = new Dictionary<string, List<BaseData>>(100);
+        private static Queue<string> CacheKeys = new Queue<string>(100);
 
         /// <summary>
         /// Event fired when the specified source is considered invalid, this may
@@ -114,7 +113,7 @@ namespace QuantConnect.Lean.Engine.DataFeeds
             }
             if (cache == null)
             {
-                cache = new List<BaseData>();
+                cache = _shouldCacheDataPoints ? new List<BaseData>(30000) : null;
                 using (var reader = CreateStreamReader(source))
                 {
                     // if the reader doesn't have data then we're done with this subscription
@@ -175,47 +174,34 @@ namespace QuantConnect.Lean.Engine.DataFeeds
                     yield break;
                 }
 
-                CacheKeys.Enqueue(cacheKey);
-                if (BaseDataSourceCache.Count > CacheSize)
+                lock (CacheKeys)
                 {
-                    var orgCache = BaseDataSourceCache;
-                    // only one thread will enter here and remove the first 33% items from the cache & queue
-                    if (Monitor.TryEnter(orgCache))
-                    {
-                        try
-                        {
-                            // we create a new dictionary so we don't have to take locks and add our new item
-                            var newCache = new Dictionary<string, List<BaseData>>(BaseDataSourceCache) { [cacheKey] = cache };
+                    CacheKeys.Enqueue(cacheKey);
+                    // we create a new dictionary, so we don't have to take locks when reading, and add our new item
+                    var newCache = new Dictionary<string, List<BaseData>>(BaseDataSourceCache) { [cacheKey] = cache };
 
-                            string key;
-                            var removeCount = 0;
-                            // we remove a portion of the first entries
-                            while (CacheKeys.TryDequeue(out key) && ++removeCount < (CacheSize / 3))
-                            {
-                                newCache.Remove(key);
-                            }
-                            // update the cache instance
-                            BaseDataSourceCache = newCache;
-                        }
-                        finally
+                    if (BaseDataSourceCache.Count > CacheSize)
+                    {
+                        var removeCount = 0;
+                        // we remove a portion of the first in entries
+                        while (++removeCount < (CacheSize / 4))
                         {
-                            Monitor.Exit(orgCache);
+                            newCache.Remove(CacheKeys.Dequeue());
                         }
+                        // update the cache instance
+                        BaseDataSourceCache = newCache;
                     }
                     else
                     {
-                        BaseDataSourceCache[cacheKey] = cache;
+                        // update the cache instance
+                        BaseDataSourceCache = newCache;
                     }
-                }
-                else
-                {
-                    BaseDataSourceCache[cacheKey] = cache;
                 }
             }
 
             if (cache == null)
             {
-                throw new InvalidOperationException($"CacheItem can not be cast into expected type. Key: {cacheKey}");
+                throw new InvalidOperationException($"Cache should not be null. Key: {cacheKey}");
             }
             // Find the first data point 10 days (just in case) before the desired date
             // and subtract one item (just in case there was a time gap and data.Time is after _date)
@@ -261,6 +247,20 @@ namespace QuantConnect.Lean.Engine.DataFeeds
         {
             var handler = CreateStreamReaderError;
             if (handler != null) handler(this, new CreateStreamReaderErrorEventArgs(date, source));
+        }
+
+        /// <summary>
+        /// Set the cache size to use
+        /// </summary>
+        /// <remarks>How to size this cache: Take worst case scenario, BTCUSD hour, 60k QuoteBar entries, which are roughly 200 bytes in size -> 11 MB * CacheSize</remarks>
+        public static void SetCacheSize(int megaBytesToUse)
+        {
+            if (megaBytesToUse != 0)
+            {
+                // we take worst case scenario, each entry is 12 MB
+                CacheSize = megaBytesToUse / 12;
+                Log.Trace($"TextSubscriptionDataSourceReader.SetCacheSize(): Setting cache size to {CacheSize} items");
+            }
         }
     }
 }