--- a/Engine/Engine.cs
+++ b/Engine/Engine.cs
@@ -91,6 +91,7 @@ namespace QuantConnect.Lean.Engine
 
             try
             {
+                TextSubscriptionDataSourceReader.SetCacheSize((int) (job.RamAllocation * 0.2));
 
                 //Reset thread holders.
                 var initializeComplete = false;