<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
Context: MemoryCache isn't enforcing the cache limit size, so memory grows.
- Adding cache based on item count, which will be based on 20% of the allocated ram for the job, taking worst-case scenario to calculated estimate entry size
- Adding performance unit tests, happy and cache miss case.

## Backtests

### Cache miss case
```
Coarse 1k + Fine 20, **RANDOM** selection (forces cache miss). 6 months
```
`MASTER` -> **Memory grows**
502.11 seconds at 0k data points per second. Processing total of 201,652 data points.
501.09 seconds at 0k data points per second. Processing total of 201,119 data points.
`No cache`
869.29 seconds at 0k data points per second. Processing total of 201,857 data points.
879.22 seconds at 0k data points per second. Processing total of 203,451 data points.
`PR`
990.31 seconds at 0k data points per second. Processing total of 202,889 data points.
919.63 seconds at 0k data points per second. Processing total of 201,379 data points

### Cache happy case
```
public override void Initialize()
{
    SetStartDate(2016, 01, 01);
    SetEndDate(2017, 01, 01);
    SetCash(10000);
    _symbol = AddEquity("SPY").Symbol;
}
public override void OnEndOfDay()
{
    var dailyHistory = History(_symbol, 1, Resolution.Hour).First();
}
```
`MASTER`
3.49 seconds at 57k data points per second. Processing total of 198,224 data points.
3.49 seconds at 57k data points per second. Processing total of 198,224 data points.
`NoCache`
31.57 seconds at 6k data points per second. Processing total of 198,224 data points
32.37 seconds at 6k data points per second. Processing total of 198,224 data points.
`PR`
4.89 seconds at 41k data points per second. Processing total of 198,224 data points.
5.50 seconds at 36k data points per second. Processing total of 198,224 data points.

## Unit tests
### Cache happy case
`Master`
Took 12396ms. Data count 58432500
Took 11907ms. Data count 58432500
`No Cache`
Took 20407ms. Data count 58432500
Took 20890ms. Data count 58432500
`PR`
11274ms. Data count 58432500
10545ms. Data count 58432500

### Cache miss case
`Master`
Took 10373ms. Data count 19477500
Took2 956ms. Data count 38955000 -> **Memory grows** not enforcing any real limits, so it's not a cache miss
`No Cache`
Took 7030ms. Data count 19477500
Took2 6868ms. Data count 38955000
`PR`
Took 9500ms. Data count 19477500
Took2 9603ms. Data count 38955000


#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
N/A
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Memory leak & speed performance
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit and regression tests, cloud backtests
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Refactor (non-breaking change which improves implementation)
- [x] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->